OJORISE-63 FEAT: API 연동
