+++ src/services/postSurvey.ts
@@ -1,13 +1,5 @@
 import api from "@/lib/axios";
-
-export interface SurveyRequest {
-  birthdate: string;
-  telecomProvider: string;
-  planName: string;
-  planPrice: number;
-  familyBundle: string;
-  familyNum: string;
-}
+import { SurveyRequest } from "@/types/survey";
 
 export async function postSurvey(data: SurveyRequest) {
   const response = await api.post("/survey", data);
