+++ src/types/survey.ts
@@ -1,0 +1,8 @@
+export interface SurveyRequest {
+  birthdate: string;
+  telecomProvider: string;
+  planName: string;
+  planPrice: number;
+  familyBundle: string;
+  familyNum: string;
+}
